+++ frontend/src/lib/state/application-slice.ts
